changed datatype records into size_t
git-svn-id: http://svn.xp-dev.com/svn/wxextension@1257 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/src/report/textfile.cpp
@@ -39,7 +39,7 @@ public:
     : m_db(db)
     , m_TextFile(file)
     , m_Records(0){}
-  int GetRecords() const {return m_Records;};
+  size_t GetRecords() const {return m_Records;};
   bool ExecQuery(const wxString& query);
   static const wxString QueryRunTimeText() {return "QUERY RUN TIME: ";};
 private:
@@ -48,7 +48,7 @@ private:
   otl_stream m_stream;
   wxExTextFile* m_TextFile;
   wxString m_Query;
-  int m_Records;
+  size_t m_Records;
 };
 #endif
